requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
openai==1.65.3
Pillow==10.1.0
requests==2.31.0
pdfkit==1.0.0
gTTS==2.3.2
